requirements_dev_py3.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements_dev_py3.txt requirements_dev_py3.in
#
entrypoints==0.3          # via flake8
flake8==3.7.9
grpcio-tools==1.26.0
grpcio==1.26.0            # via grpcio-tools
mccabe==0.6.1             # via flake8
mypy-extensions==0.4.3    # via mypy
mypy==0.910 ; python_version >= "3"
nose==1.3.7
protobuf==3.11.3          # via grpcio-tools
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
six==1.14.0               # via grpcio, protobuf
sqlalchemy-stubs==0.4 ; python_version >= "3"
toml==0.10.2              # via mypy
typed-ast==1.4.1          # via mypy
types-click==7.1.5 ; python_version >= "3"
types-flask==1.1.3 ; python_version >= "3"
types-itsdangerous==1.1.3 ; python_version >= "3"
types-jinja2==2.11.6      # via types-flask
types-markupsafe==1.1.6   # via types-jinja2
types-pytz==2021.1.2 ; python_version >= "3"
types-redis==3.5.6 ; python_version >= "3"
types-requests==2.25.6 ; python_version >= "3"
types-six==0.1.9 ; python_version >= "3"
types-werkzeug==1.0.5 ; python_version >= "3"
typing-extensions==3.10.0.0  # via mypy, sqlalchemy-stubs

# The following packages are considered to be unsafe in a requirements file:
# setuptools==57.4.0        # via protobuf
